products-service/requirements.txt
fastapi[standard]
uvicorn[standard] 
pydantic 
sqlalchemy
pyjwt
graypy